+++ src/MOCK_DB_DATA/library-items/badge/badge-circular.ts
@@ -1,6 +1,6 @@
 import { LibraryItemIds } from '../library-item-ids';
 import { LibraryItem, LibraryItemType } from 'src/app/views/shared/models/library-item.model';
-import { LibraryItemConstants as CONST } from '../../library-item-constants';
+import { LibraryItemConstants as CONST } from '../library-item-constants';
 
 export const BadgeCircular: LibraryItem[] = [{
   id: LibraryItemIds.BadgeCircular,
