+++ src/MOCK_DB_DATA/library-items/buttons/btn-shortcut-link.ts
@@ -1,6 +1,6 @@
 import { LibraryItemIds } from '../library-item-ids';
 import { LibraryItem, LibraryItemType } from 'src/app/views/shared/models/library-item.model';
-import { LibraryItemConstants as CONST } from '../../library-item-constants';
+import { LibraryItemConstants as CONST } from '../library-item-constants';
 
 export const BtnShortcutLink: LibraryItem[] = [{
   id: LibraryItemIds.ButtonShortcutLink,
